+++ lua/nvimlspconfig.lua
@@ -54,7 +54,7 @@ require('lspconfig')['tsserver'].setup {
 }
 
 require('lspconfig')['rust_analyzer'].setup {
-    on_attach = on_attach,
+    -- on_attach = on_attach,
     -- flags = lsp_flags,
     -- Server-specific settings...
     settings = {