+++ lua/plugins.lua
@@ -4,9 +4,13 @@ return require('packer').startup(function(use)
   use 'wbthomason/packer.nvim'
 
   use 'neovim/nvim-lspconfig'
+
   use {
     "williamboman/mason.nvim",
   }
+
+  use 'nvim-treesitter/nvim-treesitter'
+
   use {
     "williamboman/mason-lspconfig.nvim",
   }